refactor: simplify schemas directory structure
- Move validation schemas to single validation.ts file
- Move form configurations to single forms.ts file
- Remove redundant eticket- prefixes from file names
- Delete unnecessary re-export files and duplicate structures
- Update imports across all form components and steps
- Organize by function rather than arbitrary separation
