+++ src/components/forms/steps/general-info-step.tsx
@@ -15,7 +15,6 @@ import {
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
-import { getErrorMessage } from "@/lib/utils";
 import {
   permanentAddressSchema,
   residenceCountrySchema,
@@ -24,7 +23,8 @@ import {
   postalCodeSchema,
   hasStopsSchema,
   entryOrExitSchema,
-} from "@/lib/validations/eticket-schemas";
+} from "@/lib/schemas/validation";
+import { getErrorMessage } from "@/lib/utils";
 
 interface GeneralInfoStepProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
